function_app/requirements.txt
azure-functions>=1.11.0
azure-functions-worker>=1.0.0
